Create README.md
With Erlang gaining popularity amongst developers around the world with communities and
forums, we have decided to select Erlang news as our information domain. There are few
centralised hubs of hot news - ranging in variety - and we believe we can penetrate the market with
a more centralised solution. We have chosen the following sources that already aggregate news
from their own independent sources, however they range in variety. By combining the following we
believe we can achieve a unique product, displaying various topics associated with Erlang, such as
software updates, jobs, events, tutorials, projects and more.








The purpose of the project is to develop a business intelligence (BI) system.